feat: detail form 컴포넌트 분리, Suspense 추가
